refactor: onboarding steps의 step1, 2 폴더 구조 및 네이밍 표준화
- 폴더명을 기능 중심으로 변경 (step1→houseInfo, step2→floorPlan)
- 컴포넌트명에서 step 번호 제거
