+++ src/pages/onboarding/components/steps/floorPlan/FloorPlan.css.ts
