+++ src/pages/onboarding/components/steps/floorPlan/FloorPlanList.css.ts
