+++ src/shared/components/card/floorCard/FloorCard.tsx
@@ -1,11 +1,11 @@
 import * as styles from './FloorCard.css.ts';
 
-interface FloorCardProps extends React.ComponentProps<'div'> {
+interface FloorPlanItemProps extends React.ComponentProps<'div'> {
   src: string;
   selected?: boolean;
 }
 
-const FloorCard = ({ src, selected = false }: FloorCardProps) => {
+const FloorPlanItem = ({ src, selected = false }: FloorPlanItemProps) => {
   return (
     <div className={`${styles.container} ${selected ? styles.selected : ''}`}>
       <img src={src} className={styles.floorimg} alt="카드 이미지" />
@@ -13,4 +13,4 @@ const FloorCard = ({ src, selected = false }: FloorCardProps) => {
   );
 };
 
-export default FloorCard;
+export default FloorPlanItem;
