+++ src/stories/FloorCard.stories.tsx
@@ -1,9 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import FloorCard from '@/shared/components/card/floorCard/FloorCard';
+import FloorPlanItem from '@/shared/components/card/floorCard/FloorCard';
 
-const meta: Meta<typeof FloorCard> = {
+const meta: Meta<typeof FloorPlanItem> = {
   title: 'Components/FloorCard',
-  component: FloorCard,
+  component: FloorPlanItem,
   argTypes: {
     src: {
       control: 'text',
@@ -13,7 +13,7 @@ const meta: Meta<typeof FloorCard> = {
 };
 
 export default meta;
-type Story = StoryObj<typeof FloorCard>;
+type Story = StoryObj<typeof FloorPlanItem>;
 
 export const Default: Story = {
   args: {
